src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/text/reference/ArrayReferenceBooleanTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.array.text.ArrayTextParser;
import org.openqa.selenium.WebElement;

public class ArrayReferenceBooleanTextParser extends ArrayTextParser {

    public ArrayReferenceBooleanTextParser() {
        super(Boolean.class, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Boolean.parseBoolean(element.getText());
    }

}
